+++ test/unit/roledeps_test.lua
@@ -158,6 +158,15 @@ function g.test_full_deps()
         'storage', 'role-c', 'role-d', 'role-b', 'role-a',
     })
 
+    local enabled_roles_wo_deps = roles.get_enabled_roles_without_deps({
+        ['storage'] = true,
+        ['role-a'] = true,
+        ['vshard-router'] = true,
+    })
+    t.assert_equals(enabled_roles_wo_deps, {
+        'role-a', 'storage', 'vshard-router',
+    })
+
     local enabled_roles = roles.get_enabled_roles({
         ['vshard-storage'] = false,
         ['storage'] = true,